requirements.txt
python-dotenv
pandas
numpy
scipy
pymongo[srv]
certifi
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart
